requirements.txt
python-telegram-bot==20.3
sqlalchemy==2.0.19
requests==2.31.0
beautifulsoup4==4.12.2
flask==2.3.2
pyngrok==7.0.0
schedule==1.2.0
python-dotenv==1.0.0
PyYAML==6.0.1